create models and update migrations

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,8 +20,8 @@ class CreateUsersTable extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
-            $table->integer('role')->default('1');
-            $table->integer('ref_id')->nullable();
+            $table->bigInteger('role')->default('1');
+            $table->bigInteger('ref_id')->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
